add 'url' argument to get_instance_metadata, get_instance_userdata
This adds a 'url' arg to get_instance_metadata and get_instance_userdata

The reason for this argument is primarily to support using boto.utils in
a AWS like environment where there is a metadata service, but it is not
available in the standard location (under http://169.254.169.254/)

One such environment is Eucalyptus running in SYSTEM or STATIC modes.

Per http://open.eucalyptus.com/wiki/EucalyptusNetworkConfiguration_v2.0
| If your system is configured using SYSTEM or STATIC networking modes,
| then retrieving data requires the IP (or hostname) and port number of
| the Eucalyptus Cloud Controller (CLC):
|
| http://<IP or hostname of CLC>:8773/<specific request>

Cloud-init is addressing this under the bug at
http://bugs.launchpad.net/bugs/761847

Also thanks to Kieran Evans.