More targeted fix for content-length in POST requests.
My previous attempt messed up PUT requests to S3, because the
content-length is set by the caller, and the body is sent later.

Now, the content-length for signed POST requests is set at the point
where the body of the request is computed, and will be reset on a
retry.