New connection pool algorithm.
I have been seeing problems where AWS would time out a connection
right before it was pulled from the connection pool, and then a
variety of errors can happen, some of which boto catches, and some of
which it doesn't.  The biggest problem, though, was that it sometimes
takes a long time to get those errors.

The connection pool will now discard connections that are 'stale',
meaning that they have been in the pool for too long.  The threshold
for staleness is currently set at 15 seconds.

The test_pool program now uses real threads, not greenlets.