Fix create_scheduled_group_action to support setting to 0.
When deciding whether to pass parameter x up to AWS, the code was
using 'if x' to test whether it had been changed from the default of
None.  This didn't work for 0, so I wasn't able to shut down an
auto-scaling group by setting the number of instances to 0.

The code now uses 'if x is not None' to check whether a parameter has
been set.