Fix bucket.list() which can repeat the same keys when delimiter is used.
bucket.list will repeat the same keys multiple times when the following is true:
* delimiter is set.
* returned set is truncated.
* returned set contains common prefixes.

A typical example is:
  "a/", "c0000", "c0001", ..., "c1000"

As common prefixes are listed last by AWS, prefix "a/"
will be listed last. The first loop will return 1000 keys
of which the last is "a/". Therefore the next loop will
start listing keys "c0000" onwards thus repeating those
set of keys twice. Amazon returns the NextMarker to help
avoid this but its only available when the set is
truncated and a delimiter was used.