Fixing a bug in Layer2.update_throughput
The update_throughput method of Layer2 calls the update_from_response
method of Table, but instead of passing in the full response (as all
other similar calls to update_from_response do in the codebase) it
passes in response['TableDescription']. Since update_from_response
looks for either the key 'Table' or 'TableDescription' in the response
object, neither is found and, therefore, the table's local attributes
are not updated and do not correctly reflect the changes that have
taken place on the server. This fix passes in the full response object
to update_from_response.