add vpc_id to VPC-linked group obj before returning
This commit fixes a bug where vpc_id attribute is not set on a newly
created group object when using vpc_id kwarg of create_security_group().
This breaks all of the authorize, revoke, and delete security group
convenience methods for VPC-linked groups. A separate call to
get_all_security_groups() to re-fetch the group correctly includes the
vpc_id but this fix ensures that the object immediately returned after
creating a VPC-linked group has the proper vpc_id attribute.