s3: allow empty metadata dict in bucket.copy_key()
Previously the metadata would only be REPLACE'd if metadata was not None
AND not empty. Updated the code to only require that metadata is not
None in order to REPLACE the metadata. This allows an empty dictionary
to be used which will completely clear the metadata on the new key.

Before this change an empty metadata dict resulted in the new key's
metadata being completely copied from the original as if no metadata was
specified.