Add helper function to serialize list of structures
For query services that have operations that accept a list
of complex types, this helper function provides a simple
way to serialize those params.  There are several services
that can be refactored to use this new method, but I think
it makes sense to do that on a case by case basis (and as
separate commits).

I added unit tests for this as well, but given that these
were for the AWSQueryConnection class, I put the test file
in the top level tests/unit directory.