Update connection.py: get_path when suppress_consec_slashes is False
Should fix boto/boto#1387

When suppressing consec slashes is set to False, slashes at the beginning of the key are suppressed too aggressively.  It shouldn't turn all slashes at the beginning to empty. It should decrease the number of prefix slashes by 1.

This likely depends on self.path being ending in a "/".